SampleFieldOptions.test.tsx convert to RTL
